pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "driftpilot"
version = "0.1.0"
description = "Adaptive time series forecasting with concept drift detection"
requires-python = ">=3.11"
dependencies = [
    "lightgbm>=4.3",
    "scikit-learn>=1.4",
    "river>=0.21",
    "evidently>=0.4",
    "mlflow>=2.13",
    "pandas>=2.2",
    "pyarrow>=16",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.7",
    "streamlit>=1.35",
    "plotly>=5.22",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "requests>=2.31",
    "notebook>=7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
